+++ Tests/Decoders.swift
@@ -46,3 +46,11 @@ struct TestObjectDecoder: Decoder {
         )
     }
 }
+
+// MARK: -
+
+struct ErrorThrowingDecoder: Decoder {
+    func decode(_ object: Any) throws -> Any {
+        throw NSError(domain: "Decoder Test Error", code: 42, userInfo: nil)
+    }
+}